refactor: layoutsディレクトリをcomponentsへリネーム
分かりやすい名称とするため
